move explorer endpoints out of v0
