+++ jormungandr/src/rest/v0/mod.rs
@@ -43,7 +43,5 @@ pub fn resources() -> Vec<(
         }),
         ("/tip", &|r| r.get().with_async(handlers::get_tip)),
         ("/utxo", &|r| r.get().with_async(handlers::get_utxos)),
-        ("/graphql", &|r| r.post().with_async(handlers::graphql)),
-        ("/graphiql", &|r| r.get().with(handlers::graphiql)),
     ]
 }
